Move `pairwise`, `seq_groups`, and `is_equal` functions to `_utils.py`.
