+++ great_tables/_spanners.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 import itertools
-from typing import TYPE_CHECKING, Any, Iterable
-from collections.abc import Generator
+from typing import TYPE_CHECKING
+
 from ._gt_data import SpannerInfo, Spanners
 from ._locations import resolve_cols_c
 from ._tbl_data import SelectExpr
@@ -558,55 +558,6 @@ def empty_spanner_matrix(
         return [], vars
 
     return [{var: var for var in vars}], vars
-
-
-def pairwise(iterable: Iterable[Any]) -> Generator[tuple[Any, Any], None, None]:
-    """
-    https://docs.python.org/3/library/itertools.html#itertools.pairwise
-    pairwise('ABCDEFG') → AB BC CD DE EF FG
-    """
-    # This function can be replaced by `itertools.pairwise` if we only plan to support
-    # Python 3.10+ in the future.
-    iterator = iter(iterable)
-    a = next(iterator, None)
-    for b in iterator:
-        yield a, b
-        a = b
-
-
-def seq_groups(seq: Iterable[str]) -> Generator[tuple[str, int], None, None]:
-    iterator = iter(seq)
-
-    # TODO: 0-length sequence
-    a = next(iterator)  # will raise StopIteration if `seq` is empty
-
-    try:
-        b = next(iterator)
-    except StopIteration:
-        yield a, 1
-        return
-
-    # We can confirm that we have two elements and both are not `None`,
-    # so we can chain them back together as the original seq.
-    seq = itertools.chain([a, b], iterator)
-
-    crnt_ttl = 1
-    for crnt_el, next_el in pairwise(seq):
-        if is_equal(crnt_el, next_el):
-            crnt_ttl += 1
-        else:
-            yield crnt_el, crnt_ttl
-            crnt_ttl = 1
-
-    # final step has same elements, so we need to yield one last time
-    if is_equal(crnt_el, next_el):
-        yield crnt_el, crnt_ttl
-    else:
-        yield next_el, 1
-
-
-def is_equal(x: Any, y: Any) -> bool:
-    return x is not None and x == y
 
 
 def cols_width(data: GTSelf, cases: dict[str, str]) -> GTSelf:
